Исправил ошибки

+++ src/ru/job4j/condition/MultipleSwitchWeek.java
@@ -1,16 +1,16 @@
 package ru.job4j.condition;
 
 public class MultipleSwitchWeek {
-    public static byte numberOfDay(String name) {
-        byte value = switch (name) {
-            case "Monday", "Понедельник" -> (byte) 1;
-            case "Tuesday", "Вторник" -> (byte) 2;
-            case "Wednesday", "Среда" -> (byte) 3;
-            case "Thursday", "Четверг" -> (byte) 4;
-            case "Friday", "Пятница" -> (byte) 5;
-            case "Saturday", "Суббота" -> (byte) 6;
-            case "Sunday", "Воскресенье" -> (byte) 7;
-            default -> (byte) -1;
+    public static int numberOfDay(String name) {
+        int value = switch (name) {
+            case "Monday", "Понедельник" -> 1;
+            case "Tuesday", "Вторник" -> 2;
+            case "Wednesday", "Среда" -> 3;
+            case "Thursday", "Четверг" -> 4;
+            case "Friday", "Пятница" -> 5;
+            case "Saturday", "Суббота" -> 6;
+            case "Sunday", "Воскресенье" -> 7;
+            default -> -1;
         };
         return value;
     }
